refactor: move camunda document serializing for ai agent tools response to document package
